Dealing with #93, since it leads to an error with odmf make-db
